Refactor model command and schemas; enhance runtime presets handling
- Removed the `build_runtime_presets` function and replaced it with a call to `RuntimePresets::from_json_options` in `command.rs`.
- Added a new `config_json` field to `AgentThreadResponse` in `agent.rs` and updated the corresponding `From` implementation.
- Simplified the `From` implementation for `DomainRuntimePresets` in `models.rs` to use a constructor method.
- Changed the `fps` field type from `f32` to `f64` in `video.rs` and updated related functions and default values.
- Updated API definitions in `v1.d.ts` to reflect changes in `fps` type and added `config_json` to `AgentThreadResponse`.
- Modified test cases to initialize `config_json` in `assistant-agent-state.test.ts`.
- Enhanced the Python SDK model for `AgentThreadResponse` to include `config_json`.
- Created a new SQL migration to enforce storage contract hardening, including JSON validation and table restructuring.

+++ crates/slab-app-core/src/domain/models/video.rs
@@ -15,7 +15,7 @@ pub struct VideoGenerationCommand {
     pub width: u32,
     pub height: u32,
     pub video_frames: i32,
-    pub fps: f32,
+    pub fps: f64,
     pub cfg_scale: Option<f32>,
     pub guidance: Option<f32>,
     pub steps: Option<i32>,
